blood-report-ai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.0

# Supabase
supabase==2.4.6

# OpenAI
openai==1.35.7

# PDF extraction
PyMuPDF==1.23.26

# Image OCR
pytesseract==0.3.10
Pillow==10.3.0

# Encryption
cryptography==42.0.8

# HTTP client
httpx==0.27.0

# Data processing
numpy==1.26.4
scipy==1.13.1

# Voice (optional)
elevenlabs==1.2.2

# RAG / Vector
tiktoken==0.7.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
structlog==24.2.0